refactor(frontalier-suisse): objectifs de revenus autonomes
